+++ src/main/java/com/example/adambackend/payload/response/JwtResponse.java
@@ -7,13 +7,13 @@ import lombok.Data;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     private String roles;
 
 
-    public JwtResponse(String accessToken, Long id, String username, String email, String roles) {
+    public JwtResponse(String accessToken, Integer id, String username, String email, String roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
