+++ src/main/java/com/example/adambackend/repository/DetailProductRepository.java
@@ -8,7 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DetailProductRepository extends JpaRepository<DetailProduct, Long> {
+public interface DetailProductRepository extends JpaRepository<DetailProduct, Integer> {
     @Query("select dp from DetailProduct  dp join Product p on dp.product.id=p.id where p.id=?1")
-    List<DetailProduct> findAllByProductId(Long id);
+    List<DetailProduct> findAllByProductId(Integer id);
+
 }
