+++ src/main/java/com/example/adambackend/repository/MaterialRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialRepository  extends JpaRepository<Material, Long> {
+public interface MaterialRepository  extends JpaRepository<Material, Integer> {
 }
